removed filter builder class and made use get filted quiz query standalone
