--- src/GraphQL/GetFilteredQuiz/useGetFilteredQuiz.ts
@@ -1,4 +1,0 @@
-import { FilterConcreteBuilder, IFilterBuilder } from "./FilterBuilder";
-
-const filterBuilder: IFilterBuilder = new FilterConcreteBuilder();
-export {};
